py/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu113
matplotlib
numpy
opencv_python
Pillow
pynrrd
sahi==0.10.7
scikit_learn
scipy
tifffile
torch>=1.9.0
torchvision>=0.12.0
yolov5
pyqt6
napari
SimpleITK